requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
faiss-cpu
sentence-transformers
streamlit
requests
python-multipart
google-generativeai
python-dotenv
